Rename module from SamplingMeasurement to sampling_measurement
